+++ meta_package_manager/labels.py
@@ -104,7 +104,7 @@ MANAGER_LABEL_GROUPS: TLabelGroup = {
     "dnf-based": frozenset({"dnf", "dnf5", "yum"}),
     "dpkg-based": frozenset({"apt", "apt-mint", "opkg"}),
     "homebrew": frozenset({"brew", "cask"}),
-    "npm-based": frozenset({"npm", "yarn"}),
+    "npm-based": frozenset({"npm", "yarn", "yarn-berry"}),
     "pacman-based": frozenset({"pacman", "pacaur", "paru", "yay"}),
     "pip-based": frozenset({"pip", "pipx"}),
     "uv-based": frozenset({"uv", "uvx"}),